+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## dbt-databricks 1.11.5 (TBD)
+## dbt-databricks 1.12.0 (TBD)
+
+### Features
+
+- Enable Notebook scoped python packages installation
+
+## dbt-databricks 1.11.5 (Feb 19, 2026)
 
 ### Fixes